LPD-39341 add test
